+++ Market/MarketDay/Patches.cs
@@ -151,6 +151,7 @@ namespace MarketDay
 
     [HarmonyPatch(typeof(PathFindController))]
     [HarmonyPatch("findPathForNPCSchedules")]
+    //  alter paths through Town to travel via market shops 
     public class Postfix_findPathForNPCSchedules
     {
         public static void Postfix(Point startPoint, Point endPoint, GameLocation location, int limit, ref Stack<Point> __result)
@@ -159,39 +160,53 @@ namespace MarketDay
             if (!MarketDay.IsMarketDay()) return;
             if (!MarketDay.Config.NPCVisitors) return;
             if (MapUtility.ShopTiles.Count == 0) return;
+            
+            // if we're doing rescheduling we'll bend the pathfinding in pathfindToNextScheduleLocation
+            if (MarketDay.Config.NPCRescheduling) return; 
 
-            MarketDay.Log($"findPathForNPCSchedules {location.Name} {startPoint} -> {endPoint}", LogLevel.Trace, true);
+            __result = Schedule.pathFindViaGrangeShops(startPoint, endPoint, location, limit, 3*60);
+        }
+    }
 
-            var placesToVisit = new List<Point>();
-            foreach (var (shopX, shopY) in MapUtility.ShopTiles)
-            {
-                var visitPoint = new Point((int) shopX + Game1.random.Next(3), (int) shopY + 4);
-                if (Game1.random.NextDouble() < MarketDay.Config.StallVisitChance) placesToVisit.Add(visitPoint);
-            }
+    [HarmonyPatch(typeof(NPC))]
+    [HarmonyPatch("parseMasterSchedule")]
+    //  public Dictionary<int, SchedulePathDescription> parseMasterSchedule(string rawData)
+    //  rewrite NPC schedules to travel through town via market shops and visit them
+    //  or if they have a shop to stand next to, do that if time allows
+    public class Postfix_parseMasterSchedule
+    {
+        public static void Postfix(
+            NPC __instance, 
+            String rawData,
+            ref Dictionary<int, SchedulePathDescription> __result)
+        {
+            if (!MarketDay.Config.NPCRescheduling) return;
+            if (!MarketDay.IsMarketDay()) return;
+            
+            if (MapUtility.ShopTiles.Count == 0) return;
+            
+            __result = Schedule.parseMasterSchedule(__instance, rawData);
+        }
+    }
+    
+    [HarmonyPatch(typeof(NPC))]
+    [HarmonyPatch("getSchedule")]
+    //  public Dictionary<int, SchedulePathDescription> getSchedule(int dayOfMonth)
+    //  if NPC has no schedule for today, and they have a shop to stand next to,
+    //  schedule them to stand there all day
+    public class Postfix_getSchedule
+    {
+        public static void Postfix(
+            NPC __instance, 
+            int dayOfMonth,
+            ref Dictionary<int, SchedulePathDescription> __result)
+        {
+            if (!MarketDay.Config.NPCRescheduling) return;
+            if (!MarketDay.IsMarketDay()) return;
 
-            StardewValley.Utility.Shuffle(Game1.random, placesToVisit);
-            placesToVisit.Add(startPoint);
-            if (placesToVisit.Count < 2) return;
-
-            var waypoints = string.Join(", ", placesToVisit);
-            MarketDay.Log($"    Waypoints: {waypoints}", LogLevel.Trace, true);
-
-            // work backwards through the waypoints
-            __result = new Stack<Point>();
-
-            var thisEndPoint = endPoint;
-            foreach (var (wptX, wptY) in placesToVisit)
-            {
-                var thisStartPoint = new Point(wptX, wptY);
-
-                var originalPath = Schedule.findPathForNPCSchedules(thisStartPoint, thisEndPoint, location, limit);
-                if (originalPath is null || originalPath.Count == 0) continue;
-
-                var legPath = originalPath.ToList();
-                legPath.Reverse();
-                foreach (var pt in legPath) __result.Push(pt);
-                thisEndPoint = thisStartPoint;
-            }
+            if (__result is not null && __result.Count > 0) return;
+            
+            __result = Schedule.getScheduleWhenNoDefault(__instance, dayOfMonth);
         }
     }
 }